Reorganized search and navbar files
