+++ src/components/Navbar/MenuItems.js
@@ -33,10 +33,5 @@ export const MenuItems = [
     title: "Cart",
     url: "/cart",
     className: "nav-links",
-  },
-//   {
-//   title: "Sign up",
-//   url: "#",
-//   className: "nav-links-mobile",
-// },
+  }
 ];
