+++ src/components/Navbar/NavStyle.css
@@ -1,8 +1,30 @@
-.nav-menu {
-    display: none;
+nav {
+    display: flex;
+    flex-direction: row;
+    width: 100vw;
+    justify-content: space-between;
+    background-color: #FEF9C7;
 }
 
-.nav-menu.active {
+.navbar-logo {
+    margin: 0 2rem;
+    padding: 0;
+}
+
+nav li {
+    display: flex;
+    list-style: none;
+}
+
+nav a {
+    text-decoration: none;
+    color: gray;
+    padding: 1em;
+}
+
+.navbar-links {
     display: inline-flex;
+    padding: 0;
     flex-wrap: wrap;
+
 }
