+++ src/components/Navbar/Navbar.js
@@ -1,29 +1,16 @@
 import React, { Component } from "react";
 import { MenuItems } from "./MenuItems";
-import { Button } from "./Button";
 import "./NavStyle.css";
 
 class Navbar extends Component {
-  state = { clicked: false };
-
-  
-
-  handleClick = () => {
-    this.setState({ clicked: !this.state.clicked });
-  };
 
   render() {
     return (
       <nav className="NavbarItems">
         <h1 className="navbar-logo">
-          Ducks 'R' Us<i className="fab fa-react"></i>
+          Ducks 'R' Us
         </h1>
-        <div className="menu-icon" onClick={this.handleClick}>
-          <i
-            className={this.state.clicked ? "fas fa-times" : "fas fa-bars"}
-          ></i>
-        </div>
-        <ul className={this.state.clicked ? "nav-menu active" : "nav-menu"}>
+        <div className="navbar-links">
           {MenuItems.map((item, index) => {
             return (
               <li key={index}>
@@ -33,12 +20,11 @@ class Navbar extends Component {
               </li>
             );
           })}
-        </ul>
-        {/* <Button href="./cart">Cart</Button> */}
+        </div>
       </nav>
-      
+
     );
   }
 }
-  
+
 export default Navbar;
