--- src/style/theme.css
@@ -1,27 +1,0 @@
-html,
-body {
-    margin: 0;
-    padding: 0;
-}
-
-header nav {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-header nav li {
-    display: flex;
-    list-style: none;
-}
-
-header nav a {
-    text-decoration: none;
-    color: gray;
-    padding: 1em;
-}
-
-header nav ul {
-    display: inline-flex;
-    padding: 0;
-
-}
